fixed fp_arith and fp_conv_3_3

+++ src/CodeGen_Clockwork_Target.cpp
@@ -110,7 +110,7 @@ string type_to_c_type(Type type) {
     } else if (type.bits() == 16 && !type.is_bfloat()) {
       oss << "float16_t";
     } else if (type.bits() == 16 && type.is_bfloat()) {
-      oss << "bfloat16_t";
+      oss << "uint16_t";
     } else {
       user_error << "Can't represent a float with this many bits in C: " << type << "\n";
     }
